postman tested and cors installed

+++ books-api/controllers/bookController.js
@@ -3,9 +3,44 @@ const router = express.Router();
 const db = require('../models');
 const Book = require('../models/books');
 
+
+router.get('/seed', (req, res) => {
+    Book.insertMany([{
+        "title": "The Shinobi Initiative",
+        "description": "The reality-bending adventures of a clandestine service agency in the year 2166",
+        "year": 2014,
+        "quantity": 10,
+        "imageURL": "https://imgur.com/LEqsHy5.jpeg"
+      },
+      {
+        "title": "Tess the Wonder Dog",
+        "description": "The tale of a dog who gets super powers",
+        "year": 2007,
+        "quantity": 3,
+        "imageURL": "https://imgur.com/cEJmGKV.jpg"
+      },
+      {
+        "title": "The Annals of Arathrae",
+        "description": "This anthology tells the intertwined narratives of six fairy tales.",
+        "year": 2016,
+        "quantity": 8,
+        "imageURL": "https://imgur.com/VGyUtrr.jpeg"
+      },
+      {
+        "title": "Wâˆ€RP",
+        "description": "A time-space anomaly folds matter from different points in earth's history in on itself, sending six unlikely heroes on a race against time as worlds literally collide.",
+        "year": 2010,
+        "quantity": 4,
+        "imageURL": "https://imgur.com/qYLKtPH.jpeg"
+      }])
+        .then(() => res.status(200).json({ message: 'Seed successful '}))
+        .catch(err => res.status(400).json({ message: 'Seed unsuccessful', error: err.message }))
+});
+
+
 router.get('/', async (req, res) => {
     try {
-        const books = await db.Book.find([]);
+        const books = await Book.find();
         res.status(200).json(books);
     } 
     catch (err)
@@ -16,7 +51,7 @@ router.get('/', async (req, res) => {
 
 router.get('/:id', async (req, res) => {
     try {
-        const book = await db.Book.findById(req.params.id);
+        const book = await Book.findById(req.params.id);
         if (!book) throw new Error('no book found');
         res.status(200).json(book);
     }
@@ -31,16 +66,27 @@ router.post('/', (req, res) => {
         .catch(err => res.status(500).json({ error: err.message }));
 });
 
-router.put('/:id', (req, res) => {
-    Book.findByIdAndUpdate(req.params.id, req.body, { new: true})
-    .then(updateBook => res.status(200).json(updatedBook))
-    .catch(err => res.status(500).json({ error: err.message }));
+router.put('/:id', async (req, res) => {
+    try {
+        const book = await Book.findById(req.params.id);
+        if (!book) throw new Error('No book found');
+        const updateBook = await Book.findByIdAndUpdate(req.params.id, req.body, { new: true});
+        res.status(200).json(updateBook);
+    } catch (err) {
+        res.status(500).json({ error: err.message });
+
+    }
 });
 
-router.delete('/:id', (req, res) => {
-    Book.findByIdAndDelete(req.params.id)
-    .then(() => res.status(200).json({ message: 'Book deleted successfully'}))
-    .catch(err => res.status(500).json({ error: err.message }));
+router.delete('/:id', async (req, res) => {
+    try {
+        const book = await Book.findById(req.params.id);
+        if (!book) throw new Error('No book found');
+        await Book.findByIdAndDelete(req.params.id);
+        res.status(200).json({ message: 'Book deleted successfully'});
+    } catch (err) {
+        res.status(500).json({ error: err.message });
+    }
 });
 
 module.exports = router;
